Restrict deletion of serial when screenings are still attached

+++ app/Http/Controllers/SerialController.php
@@ -81,7 +81,10 @@ class SerialController extends Controller
         if (Auth::user()->level < Config::get('constants.auth_level.editor')) {
             abort(403);
         }
-        $serial = Serial::firstWhere('uuid', $request->uuid);
+        $serial = Serial::where('uuid', $request->uuid)->with('screenings')->first();
+        if (count($serial->screenings) > 0) {
+            abort(409, 'Die Reihe kann nicht gelöscht werden, solange ihr noch Vorführungen zugeordnet sind.');
+        }
         $image = $serial->image;
 
         $serial->delete();
